Refactor project structure: rename modules to 'search-engine-backend' and 'search-engine-events'
