+++ search-engine-backend/src/main/resources/templates/recommendForUserQuery.html
