+++ search-engine-backend/src/test/java/de/uni_leipzig/search_engine/backend/controller/SearchControllerLinksTest.checkEmptyResultAsInput.approved.txt
