+++ search-engine-backend/src/test/java/de/uni_leipzig/search_engine/backend/controller/SearchControllerTest.approveEmptyQueryInput.approved.txt
