--- search-engine-uni-backend/src/main/java/de/uni_leipzig/search_engine_uni/backend/aspects/dto/SearchResponseEvent.java
@@ -1,15 +1,0 @@
-package de.uni_leipzig.search_engine_uni.backend.aspects.dto;
-
-import groovy.transform.EqualsAndHashCode;
-import lombok.Data;
-import lombok.experimental.Accessors;
-
-@Data
-@Accessors(chain=true)
-@EqualsAndHashCode(callSuper=true)
-public class SearchResponseEvent extends WebResponseEvent
-{
-	private String query;
-	
-	private int resultPage;
-}
