--- search-engine-uni-backend/src/main/java/de/uni_leipzig/search_engine_uni/backend/aspects/dto/SearchResultSelectedEvent.java
@@ -1,19 +1,0 @@
-package de.uni_leipzig.search_engine_uni.backend.aspects.dto;
-
-import groovy.transform.EqualsAndHashCode;
-import lombok.Data;
-import lombok.experimental.Accessors;
-
-@Data
-@Accessors(chain=true)
-@EqualsAndHashCode(callSuper=true)
-public class SearchResultSelectedEvent extends WebResponseEvent
-{
-	private String query;
-	
-	private int searchPage = 1;
-	
-	private int resultId;
-	
-	private String documentUri;
-}
